Extract :wrench: NPC movement logic into new pathfinding module.
